read the wiki

+++ Matrix/display_matrix.cpp
@@ -3,9 +3,9 @@
 #include "display_matrix.h"
 using namespace std;
 
-void displayMatrix(const std::vector<std::vector<int>>& matrix){
+void displayMatrix(const std::vector<std::vector<double>>& matrix){
 	for (const auto& row : matrix) {
-		for (int num : row) {
+		for (double num : row) {
 			cout << num << " ";
 		}
 		cout << endl;
